[python-fastapi] support pydantic v2 models
Previously, the generated `additional_properties` field showed up within
the response of the generated API as opposed marshaling the model so that its fields are added to the root object.
Apparently that is because pydantic v2 does not honour the generated `to_dict` methods anymore (which would have mapped the object to the correct representation) but, instead, supports additional properties natively by specifying `extra=allow` within the `model_config`.

Correspondingly, the following changes have been applied:
* To allow additional fields, specify `extra=allow` within the `model_config`.
* Don't generate the `additional_properties` field - users can use pydantic's built-in `model.extra_fields` instead.
* Don't generate the `{to|from}_{dict|json}` methods since pydantic is taking care of the model mapping based on the declared fields and `model_config` - users can use `model.model_dump[_json]` instead.